# Add polylines and spheres for 