--- a/rmf_site_editor/src/interaction/anchor.rs
+++ b/rmf_site_editor/src/interaction/anchor.rs
@@ -20,7 +20,7 @@ use crate::{
     interaction::IntersectGroundPlaneParams,
     interaction::*,
     keyboard::DebugMode,
-    site::{Anchor, Category, Delete, Dependents, SiteAssets, Subordinate},
+    site::{Anchor, Category, Delete, Dependents, SiteAssets, Subordinate, PointAsset},
 };
 use bevy::prelude::*;
 
@@ -38,6 +38,7 @@ pub fn add_anchor_visual_cues(
     >,
     categories: Query<&Category>,
     site_assets: Res<SiteAssets>,
+    point_assets: Res<PointAsset>,
     interaction_assets: Res<InteractionAssets>,
 ) {
     for (e, parent, subordinate, anchor) in &new_anchors {
@@ -55,6 +56,11 @@ pub fn add_anchor_visual_cues(
                 ..default()
             });
             body.insert(Selectable::new(e));
+            body.insert(MaterialMeshBundle {
+                mesh: point_assets.bevy_point_mesh.clone(),
+                material: point_assets.bevy_point_material.clone(),
+                ..default()
+            });
             if subordinate.is_none() {
                 body.insert(DragPlaneBundle::new(e, Vec3::Z));
             }