--- a/rmf_site_editor/src/interaction/camera_controls.rs
+++ b/rmf_site_editor/src/interaction/camera_controls.rs
@@ -49,9 +49,12 @@ pub const HOVERED_OUTLINE_LAYER: u8 = 4;
 /// The X-Ray layer is used to show visual cues that need to be rendered
 /// above anything that would be obstructing them.
 pub const XRAY_RENDER_LAYER: u8 = 5;
-/// The Picking layer contains an entity based color map for picking entities
-/// which are drawn in screen space.
-pub const PICKING_LAYER: u8 = 6;
+/// The Line Picking layer contains an entity based color map for picking
+/// polylines which are drawn in screen space.
+pub const LINE_PICKING_LAYER: u8 = 6;
+/// The Line Picking layer contains an entity based color map for picking
+/// points which are drawn in screen space.
+pub const POINT_PICKING_LAYER: u8 = 7;
 
 #[derive(Resource)]
 pub struct MouseLocation {