--- a/rmf_site_editor/src/site/assets.rs
+++ b/rmf_site_editor/src/site/assets.rs
@@ -17,6 +17,7 @@
 
 use crate::{shapes::*, site::*};
 use bevy::{math::Affine3A, prelude::*};
+use bevy_points::prelude::{PointsMaterial, PointsMesh};
 
 #[derive(Resource)]
 pub struct SiteAssets {
@@ -50,6 +51,7 @@ pub struct SiteAssets {
     pub physical_camera_material: Handle<StandardMaterial>,
     pub occupied_material: Handle<StandardMaterial>,
     pub default_mesh_grey_material: Handle<StandardMaterial>,
+
 }
 
 impl FromWorld for SiteAssets {
@@ -233,6 +235,7 @@ impl FromWorld for SiteAssets {
             physical_camera_material,
             occupied_material,
             default_mesh_grey_material,
+            
         }
     }
 }
@@ -247,6 +250,36 @@ impl SiteAssets {
             &self.passive_anchor_material
         } else {
             &self.unassigned_anchor_material
+        }
+    }
+}
+
+#[derive(Resource)]
+pub struct PointAsset {
+    pub bevy_point_material: Handle<PointsMaterial>,
+    pub bevy_point_mesh: Handle<Mesh>,
+}
+
+impl FromWorld for PointAsset {
+    fn from_world(world: &mut World) -> Self {
+        let mut points_materials = world
+            .get_resource_mut::<Assets<PointsMaterial>>()
+            .unwrap();
+        let bevy_point_material = points_materials.add(PointsMaterial {
+            point_size: 0.5,       // Defines the size of the points. 
+            perspective: true,      // Specify whether points' size is attenuated by the camera depth. 
+            circle: true,           // Specify whether the shape of points is circular or rectangular.
+            ..Default::default()
+        });
+
+        let mut meshes = world.get_resource_mut::<Assets<Mesh>>().unwrap();
+        let bevy_point_mesh = meshes.add(PointsMesh::from_iter((0..1).map(|i| {
+            Vec3::ZERO
+        })).into());
+
+        Self {
+            bevy_point_material,
+            bevy_point_mesh
         }
     }
 }