--- a/rmf_site_editor/src/site/lane.rs
+++ b/rmf_site_editor/src/site/lane.rs
@@ -39,6 +39,7 @@ pub struct LaneSegments {
     pub mid: Entity,
     pub end: Entity,
     pub polyline: Entity,
+    pub picker: Entity,
     pub outlines: [Entity; 3],
 }
 
@@ -124,7 +125,7 @@ pub fn add_lane_visuals(
             .point_in_parent_frame_of(edge.end(), Category::Lane, e)
             .unwrap();
         let mut commands = commands.entity(e);
-        let (layer, start, mid, end, polyline, outlines) = commands.add_children(|parent| {
+        let (layer, start, mid, end, polyline, picker, outlines) = commands.add_children(|parent| {
             // Create a "layer" entity that manages the height of the lane,
             // determined by the DisplayHeight of the graph.
             let mut layer_cmd = parent.spawn(SpatialBundle {
@@ -132,7 +133,7 @@ pub fn add_lane_visuals(
                 ..default()
             });
 
-            let (start, mid, end, polyline, outlines) = layer_cmd.add_children(|parent| {
+            let (start, mid, end, polyline, picker, outlines) = layer_cmd.add_children(|parent| {
                 let mut start = parent.spawn(PbrBundle {
                     mesh: assets.lane_end_mesh.clone(),
                     material: lane_material.clone(),
@@ -201,17 +202,20 @@ pub fn add_lane_visuals(
                     })
                     .insert(Selectable::new(e))
                     .id();
+                let picker = parent.spawn(
+                        ScreenSpaceSelection::polyline(start_anchor, end_anchor, 10.0)).id();
 
                 (
                     start,
                     mid,
                     end,
                     polyline,
+                    picker,
                     [start_outline, mid_outline, end_outline],
                 )
             });
 
-            (layer_cmd.id(), start, mid, end, polyline, outlines)
+            (layer_cmd.id(), start, mid, end, polyline, picker, outlines)
         });
 
         commands
@@ -221,6 +225,7 @@ pub fn add_lane_visuals(
                 mid,
                 end,
                 polyline,
+                picker,
                 outlines,
             })
             .insert(SpatialBundle {
@@ -241,6 +246,7 @@ fn update_lane_visuals(
     transforms: &mut Query<&mut Transform>,
     polylines: &Query<&Handle<Polyline>>,
     polyline_assets: &mut Assets<Polyline>,
+    picker: &mut Query<&mut ScreenSpaceSelection>,
 ) {
     let start_anchor = anchors
         .point_in_parent_frame_of(edge.left(), Category::Lane, entity)
@@ -261,6 +267,19 @@ fn update_lane_visuals(
     if let Ok(polyline) = polylines.get(segments.polyline) {
         if let Some(mut polyline) = polyline_assets.get_mut(polyline) {
             polyline.vertices = vec![start_anchor, end_anchor];
+        }
+    }
+
+    if let Ok(polyline) = polylines.get(segments.polyline) {
+        if let Some(mut polyline) = polyline_assets.get_mut(polyline) {
+            polyline.vertices = vec![start_anchor, end_anchor];
+        }
+    }
+
+    if let Ok(mut picker) = picker.get_mut(segments.picker) {
+        if let ScreenSpaceSelection::Polyline(ply) = picker.as_mut() {
+            ply.start = start_anchor;
+            ply.end = end_anchor;
         }
     }
 }
@@ -284,6 +303,7 @@ pub fn update_changed_lane(
     polylines: Query<&Handle<Polyline>>,
     mut polyline_assets: ResMut<Assets<Polyline>>,
     current_level: Res<CurrentLevel>,
+    mut picker: Query<&mut ScreenSpaceSelection>,
 ) {
     for (e, edge, associated, segments, mut visibility) in &mut lanes {
         update_lane_visuals(
@@ -294,6 +314,7 @@ pub fn update_changed_lane(
             &mut transforms,
             &polylines,
             &mut polyline_assets,
+            &mut picker
         );
 
         let is_visible =
@@ -317,6 +338,7 @@ pub fn update_lane_for_moved_anchor(
     mut transforms: Query<&mut Transform>,
     polylines: Query<&Handle<Polyline>>,
     mut polyline_assets: ResMut<Assets<Polyline>>,
+    mut picker: Query<&mut ScreenSpaceSelection>,
 ) {
     for dependents in &changed_anchors {
         for dependent in dependents.iter() {
@@ -329,6 +351,7 @@ pub fn update_lane_for_moved_anchor(
                     &mut transforms,
                     &polylines,
                     &mut polyline_assets,
+                    &mut picker
                 );
             }
         }