Addresses #134 

This PR adds fixed-sized meshes for points and lanes. It also adds the logic required to select objects at a large distance. This means no matter how far away you are, you will always be able to manipulate anchors and lanes.