@xiyuoh 2023-08-04
I'm trying this out briefly with the demo world, and I'm finding it a little difficult to manipulate anchors when zooming out further. When I'm selecting an anchor, it may only allow me to move another anchor.

https://github.com/open-rmf/rmf_site/assets/30265743/e9ccd4f1-d26b-4518-adf2-bb16ac2479e7