@luca-della-vedova 2023-10-06
I was worried that doing this copying back and forth from CPU buffers might impact performance but I decided to quantify the performance to have some proper data.
For reference, I have a laptop with a `11th Gen Intel(R) Core(TM) i7-1185G7` with integrated GPU.

I used the instructions in the [cheatbook](https://bevy-cheatbook.github.io/cookbook/print-framerate.html) to add a printout of the app FPS and ran the demo world.
Before the PR:

```
2023-10-06T08:35:23.399101Z  INFO bevy diagnostic: frame_count                     : 529.000000
2023-10-06T08:35:24.387194Z  INFO bevy diagnostic: frame_time                      :   25.348438ms (avg 25.126539ms)
2023-10-06T08:35:24.387227Z  INFO bevy diagnostic: fps                             :   39.457123   (avg 39.808914)
```

After the PR:

```
2023-10-06T08:42:15.563158Z  INFO bevy diagnostic: frame_count                     : 524.000000
2023-10-06T08:42:16.502928Z  INFO bevy diagnostic: frame_time                      :  119.953480ms (avg 136.121408ms)
2023-10-06T08:42:16.502951Z  INFO bevy diagnostic: fps                             :    8.336565   (avg 7.384954)
```

It seems the FPS dropped by approximately 80%.
I have some other feedback about the UX but thought this is probably a large enough issue that would make it hard to bring this to `main` unless it's either optimized a fair bit or hidden behind a feature flag / default off.
What do you think?